Create a Consensus Adjacency Matrix from Multiple Networks

Builds a consensus adjacency matrix from a list of networks using one of three methods: "vote", "union", or "INet".

Usage

create_consensus(
  adj_matrix_list,
  method = "vote",
  weighted_list = NULL,
  theta = 0.04,
  threshold = 0.5,
  ncores = 1
)

Arguments

adj_matrix_list

A list of binary adjacency matrices (square, 0/1) with identical dimensions and matching row/column names.

method

Character string specifying the consensus strategy. One of:

"vote" (default): An edge is included if supported by at least threshold fraction of matrices.

"union": An edge is included if present in any matrix.

"INet": Combines normalized weighted matrices using consensusNet.

weighted_list

A list of weighted adjacency matrices (required if method = "INet").

theta

Numeric. Tuning parameter passed to consensusNet (default: 0.04).

threshold

Numeric between 0 and 1. Threshold for "vote" and "INet" methods. Default is 0.5.

ncores

Integer. Number of CPU cores to use when method = "INet". Default is 1.

Value

A square consensus adjacency matrix (binary or weighted, depending on the method).

Details

Consensus construction depends on the selected method:

vote

Counts the presence of each edge across all matrices and includes edges supported by at least threshold × N matrices.

union

Includes any edge that appears in any matrix.

INet

Multiplies binary matrices by corresponding weighted matrices, normalizes the results, and applies consensusNet to generate a consensus network.

For "INet", both binary and weighted adjacency matrices must be provided with matching dimensions.

Examples

data(count_matrices)

networks <- infer_networks(
    count_matrices_list = count_matrices,
    method = "GENIE3",
    nCores = 1
)
head(networks[[1]])
#>   regulatoryGene targetGene    weight
#> 1          EIF3E      PCBP1 0.1469479
#> 2          PTPRC       CYBA 0.1414658
#> 3          EEF1G      PTPRC 0.1412760
#> 4         EEF1A1       NACA 0.1327209
#> 5           CD3E       CD3D 0.1292895
#> 6         SNRPD2       CD3D 0.1274819

wadj_list <- generate_adjacency(networks)
swadj_list <- symmetrize(wadj_list, weight_function = "mean")

binary_listj <- cutoff_adjacency(
    count_matrices = count_matrices,
    weighted_adjm_list = swadj_list,
    n = 2,
    method = "GENIE3",
    quantile_threshold = 0.99,
    nCores = 1,
    debug = TRUE
)
#> [Method: GENIE3] Matrix 1 → Cutoff = 0.09218
#> [Method: GENIE3] Matrix 2 → Cutoff = 0.09540
#> [Method: GENIE3] Matrix 3 → Cutoff = 0.08970
head(binary_listj[[1]])
#>         ACTG1 ARHGDIB BTF3 CD3D CD3E CFL1 COX4I1 COX7C CYBA EEF1A1 EEF1D EEF1G
#> ACTG1       0       0    0    0    0    0      0     0    0      0     0     0
#> ARHGDIB     0       0    0    0    0    0      0     0    0      0     0     0
#> BTF3        0       0    0    0    0    0      0     0    0      0     0     0
#> CD3D        0       0    0    0    1    0      0     0    0      0     0     0
#> CD3E        0       0    0    1    0    0      0     0    0      0     0     0
#> CFL1        0       0    0    0    0    0      0     0    0      0     0     0
#>         EEF2 EIF1 EIF3E FTH1 FTL HLA-A HLA-B HLA-C HLA-E HNRNPA1 HNRNPA2B1 MYL6
#> ACTG1      0    0     0    0   0     0     0     0     0       0         0    0
#> ARHGDIB    0    0     0    0   0     0     0     0     0       0         0    0
#> BTF3       0    0     0    0   0     0     0     0     0       0         0    0
#> CD3D       0    0     0    0   0     0     0     0     0       0         0    0
#> CD3E       0    0     0    0   0     0     0     0     0       0         0    0
#> CFL1       0    0     0    0   0     0     0     0     0       0         0    0
#>         NACA PCBP1 PCBP2 PFN1 PTPRC PTPRCAP RAC2 SKP1 SNRPD2 TMSB4X UBA52 UBC
#> ACTG1      0     0     0    0     0       0    0    0      0      0     0   0
#> ARHGDIB    0     0     0    0     0       0    0    0      0      0     0   0
#> BTF3       0     0     0    0     0       0    0    0      0      0     0   0
#> CD3D       0     0     0    0     0       0    0    0      1      0     0   0
#> CD3E       0     0     0    0     0       0    0    0      0      0     0   0
#> CFL1       0     0     0    0     0       0    0    0      0      0     0   0
#>         UQCRB YBX1
#> ACTG1       0    0
#> ARHGDIB     0    0
#> BTF3        0    0
#> CD3D        0    0
#> CD3E        0    0
#> CFL1        0    0

consensus <- create_consensus(binary_listj, method = "vote")
head(consensus)
#>         ACTG1 ARHGDIB BTF3 CD3D CD3E CFL1 COX4I1 COX7C CYBA EEF1A1 EEF1D EEF1G
#> ACTG1       0       0    0    0    0    0      0     0    0      0     0     0
#> ARHGDIB     0       0    0    0    0    0      0     0    0      0     0     0
#> BTF3        0       0    0    0    0    0      0     0    0      0     0     0
#> CD3D        0       0    0    0    0    0      0     0    0      0     0     0
#> CD3E        0       0    0    0    0    0      0     0    0      0     0     0
#> CFL1        0       0    0    0    0    0      0     0    0      0     0     0
#>         EEF2 EIF1 EIF3E FTH1 FTL HLA-A HLA-B HLA-C HLA-E HNRNPA1 HNRNPA2B1 MYL6
#> ACTG1      0    0     0    0   0     0     0     0     0       0         0    0
#> ARHGDIB    0    0     0    0   0     0     0     0     0       0         0    0
#> BTF3       0    0     0    0   0     0     0     0     0       0         0    0
#> CD3D       0    0     0    0   0     0     0     0     0       0         0    0
#> CD3E       0    0     0    0   0     0     0     0     0       0         0    0
#> CFL1       0    0     0    0   0     0     0     0     0       0         0    0
#>         NACA PCBP1 PCBP2 PFN1 PTPRC PTPRCAP RAC2 SKP1 SNRPD2 TMSB4X UBA52 UBC
#> ACTG1      0     0     0    0     0       0    0    0      0      0     0   0
#> ARHGDIB    0     0     0    0     0       0    0    0      0      0     0   0
#> BTF3       0     0     0    0     0       0    0    0      0      0     0   0
#> CD3D       0     0     0    0     0       0    0    0      0      0     0   0
#> CD3E       0     0     0    0     0       0    0    0      0      0     0   0
#> CFL1       0     0     0    0     0       0    0    0      0      0     0   0
#>         UQCRB YBX1
#> ACTG1       0    0
#> ARHGDIB     0    0
#> BTF3        0    0
#> CD3D        0    0
#> CD3E        0    0
#> CFL1        0    0