Threshold Adjacency Matrices Based on Shuffled Network Quantiles

Applies a cutoff to weighted adjacency matrices using a percentile estimated from shuffled versions of the original expression matrices. Supports inference methods "GENIE3", "GRNBoost2", and "JRF".

Usage

cutoff_adjacency(
  count_matrices,
  weighted_adjm_list,
  n,
  method = "GENIE3",
  quantile_threshold = 0.99,
  weight_function = "mean",
  nCores = 1,
  grnboost_modules = NULL,
  debug = FALSE
)

Arguments

count_matrices

A list of expression matrices (genes × cells) or Seurat or SingleCellExperiment objects.

weighted_adjm_list

A list of weighted adjacency matrices (one per expression matrix) to threshold.

n

Integer. Number of shuffled replicates generated per original expression matrix.

method

Character string. One of "GENIE3", "GRNBoost2", or "JRF".

quantile_threshold

Numeric. The quantile used to define the cutoff. Default is 0.99.

weight_function

Character string or function used to symmetrize adjacency matrices ("mean", "max", etc.).

nCores

Integer. Number of CPU cores to use for parallelization. Default is the number of workers in the current BiocParallel backend.

grnboost_modules

Python modules needed for GRNBoost2 if using reticulate.

debug

Logical. If TRUE, prints detailed progress messages. Default is FALSE.

Value

A list of binary (thresholded) adjacency matrices, each corresponding to an input weighted matrix.

Details

For each input expression matrix, n shuffled versions are generated by randomly permuting each gene’s expression across cells. Network inference is performed on the shuffled matrices, and a cutoff is determined as the specified quantile (quantile_threshold) of the resulting edge weights. The original weighted adjacency matrices are then thresholded using these estimated cutoffs.

Parallelization is handled via BiocParallel.

The methods are based on:

GENIE3: Random Forest-based inference (Huynh-Thu et al., 2010).

GRNBoost2: Gradient boosting trees using arboreto (Moerman et al., 2019).

JRF: Joint Random Forests across multiple conditions (Petralia et al., 2015).

Examples

data(count_matrices)

networks <- infer_networks(
    count_matrices_list = count_matrices,
    method = "GENIE3",
    nCores = 15
)
head(networks[[1]])
#>   regulatoryGene targetGene    weight
#> 1           NACA       BTF3 0.2663768
#> 2          HLA-B     TMSB4X 0.2467107
#> 3          HLA-C       PFN1 0.2385148
#> 4          HLA-B      HLA-E 0.2013621
#> 5           NACA       EEF2 0.1956099
#> 6          ACTG1      HLA-B 0.1899475

wadj_list <- generate_adjacency(networks)
swadj_list <- symmetrize(wadj_list, weight_function = "mean")

binary_listj <- cutoff_adjacency(
    count_matrices = count_matrices,
    weighted_adjm_list = swadj_list,
    n = 2,
    method = "GENIE3",
    quantile_threshold = 0.99,
    nCores = 15,
    debug = TRUE
)
#> [Method: GENIE3] Matrix 1 → Cutoff = 0.21566
#> [Method: GENIE3] Matrix 2 → Cutoff = 0.19096
#> [Method: GENIE3] Matrix 3 → Cutoff = 0.18098
head(binary_listj[[1]])
#>        ACTG1 BTF3 CD3E CFL1 EEF1A1 EEF1D EEF1G EEF2 FTH1 FTL HLA-A HLA-B HLA-C
#> ACTG1      0    0    0    0      0     0     0    0    0   0     0     0     0
#> BTF3       0    0    0    0      0     0     0    0    0   0     0     0     0
#> CD3E       0    0    0    0      0     0     0    0    0   0     0     0     0
#> CFL1       0    0    0    0      0     0     0    0    0   0     0     0     0
#> EEF1A1     0    0    0    0      0     0     0    0    0   0     0     0     0
#> EEF1D      0    0    0    0      0     0     0    0    0   0     0     0     0
#>        HLA-E NACA PFN1 TMSB4X UBA52
#> ACTG1      0    0    0      0     0
#> BTF3       0    1    0      0     0
#> CD3E       0    0    0      0     0
#> CFL1       0    0    0      0     0
#> EEF1A1     0    0    0      0     0
#> EEF1D      0    0    0      0     0