Modify Cell Names and Combine Datasets

Modifies cell identifiers of each element in a list of matrices, Seurat objects, or SingleCellExperiment objects by appending a unique matrix index (e.g., "-m1", "-m2", etc.). After renaming, the datasets are merged into a single object.

Usage

earlyj(input_list, rowg = TRUE)

Arguments

input_list

A list of matrices, Seurat objects, or SingleCellExperiment objects. All elements must be of the same class.

rowg

Logical. If TRUE (default), genes are assumed to be rows and cells columns. If FALSE, matrices are transposed before renaming and combining.

Value

A combined matrix, Seurat object, or SingleCellExperiment object with modified (unique) cell names.

Details

For matrices, this function optionally transposes the input before combining. For Seurat and SingleCellExperiment objects, only features (genes) common across all input datasets are retained before merging. The cell names are suffixed with "-m1", "-m2", etc., according to their original list position.

Examples

data(count_matrices)

earlym <- earlyj(count_matrices)
head(earlym)
#>        cell_1-m1 cell_2-m1 cell_3-m1 cell_4-m1 cell_5-m1 cell_6-m1 cell_7-m1
#> ACTG1          7         4         0         1         4         4        10
#> BTF3           1         2         8         6         7         3         3
#> CD3E           4         9         0        11         1         2         2
#> CFL1           0        13         0         1         8         2         5
#> EEF1A1         6         2         3         4         1         1         1
#> EEF1D          4         2         3         2         1         2         1
#>        cell_8-m1 cell_9-m1 cell_10-m1 cell_11-m1 cell_12-m1 cell_13-m1
#> ACTG1          9         4          5          4          6          3
#> BTF3           3         2          4          3          0          1
#> CD3E           2         0          2          1          6          2
#> CFL1           2         1          2          0          6          2
#> EEF1A1         0         3          1          0          2         10
#> EEF1D          2         5          3          4          5          2
#>        cell_14-m1 cell_15-m1 cell_16-m1 cell_17-m1 cell_18-m1 cell_19-m1
#> ACTG1           2         10          1          3          3          6
#> BTF3            2          3          4          4          1          0
#> CD3E            2          1          5          6          2          1
#> CFL1            0          7          2          0          2          0
#> EEF1A1          4          0          1          6          3          3
#> EEF1D           1          1          1          4          2          2
#>        cell_20-m1 cell_1-m2 cell_2-m2 cell_3-m2 cell_4-m2 cell_5-m2 cell_6-m2
#> ACTG1           1         3         3         5         1         4         3
#> BTF3            0         1         1         0         1         0         2
#> CD3E            0         4         1         2         1         3         1
#> CFL1            1         0         1         0         2         1         1
#> EEF1A1          8         3         0         2         1         0         1
#> EEF1D           3         4         2         3         1         3         4
#>        cell_7-m2 cell_8-m2 cell_9-m2 cell_10-m2 cell_11-m2 cell_12-m2
#> ACTG1          1         1         5          6          4          2
#> BTF3           0         0         1          0          1          0
#> CD3E           5         2         1          1          2          2
#> CFL1           1         5         2          0          2          4
#> EEF1A1         4         0         2          6          1          1
#> EEF1D          6         4         5          4          5          1
#>        cell_13-m2 cell_14-m2 cell_15-m2 cell_16-m2 cell_17-m2 cell_18-m2
#> ACTG1           1          1          2          3          1          7
#> BTF3            5          5          1          3          0          0
#> CD3E            1          3          2          2          2          1
#> CFL1            6          6          2          2          0          1
#> EEF1A1          0          3          1          0          1          2
#> EEF1D           7          6          1          3          1          4
#>        cell_19-m2 cell_20-m2 cell_1-m3 cell_2-m3 cell_3-m3 cell_4-m3 cell_5-m3
#> ACTG1           2          4         4        10         2         2         3
#> BTF3            7          1         1         1         8         0         1
#> CD3E            1          3         6         6         1         5         0
#> CFL1            5          2         3         2         2         2         0
#> EEF1A1          1          3         8         5         2         2         4
#> EEF1D           2          5         3         2         5         8         2
#>        cell_6-m3 cell_7-m3 cell_8-m3 cell_9-m3 cell_10-m3 cell_11-m3 cell_12-m3
#> ACTG1          2         3         1         3          1          2          1
#> BTF3           1         0         0         1          1          2          2
#> CD3E           0         3         2         1          0          0          3
#> CFL1           3         4         8         3          4          4          2
#> EEF1A1         6         1         3         1          2         10          5
#> EEF1D          1         1         3         3          4          7          4
#>        cell_13-m3 cell_14-m3 cell_15-m3 cell_16-m3 cell_17-m3 cell_18-m3
#> ACTG1           2          1          4          3          3          1
#> BTF3            6          0          5          4          4          2
#> CD3E            1          1          5          5          2          0
#> CFL1            9          1          3          0          6          4
#> EEF1A1          3          7          0          0          4          3
#> EEF1D           2          6          6          0          0          7
#>        cell_19-m3 cell_20-m3
#> ACTG1           3          1
#> BTF3            4          0
#> CD3E            0          9
#> CFL1            0          1
#> EEF1A1          4          3
#> EEF1D           0          1