Modify Cell Names and Combine Datasets

Modifies cell identifiers of each element in a list of matrices, Seurat objects, or SingleCellExperiment objects by appending a unique matrix index (e.g., "-m1", "-m2", etc.). After renaming, the datasets are merged into a single object.

Usage

earlyj(input_list, rowg = TRUE)

Arguments

input_list

A list of matrices, Seurat objects, or SingleCellExperiment objects. All elements must be of the same class.

rowg

Logical. If TRUE (default), genes are assumed to be rows and cells columns. If FALSE, matrices are transposed before renaming and combining.

Value

A combined matrix, Seurat object, or SingleCellExperiment object with modified (unique) cell names.

Details

For matrices, this function optionally transposes the input before combining. For Seurat and SingleCellExperiment objects, only features (genes) common across all input datasets are retained before merging. The cell names are suffixed with "-m1", "-m2", etc., according to their original list position.

Examples

data(count_matrices)

earlym <- earlyj(count_matrices)
head(earlym)
#>         cell_1-m1 cell_2-m1 cell_3-m1 cell_4-m1 cell_5-m1 cell_6-m1 cell_7-m1
#> ACTG1           2         3         7         9         2         7         7
#> ARHGDIB         1         1         5         1         2         3         1
#> BTF3            4         2         0         2         3         2         6
#> CD3D            1         0         3         1         0         0        11
#> CD3E            2         1         3         2         2         1        11
#> CFL1            0         4         0         1         0         0         5
#>         cell_8-m1 cell_9-m1 cell_10-m1 cell_11-m1 cell_12-m1 cell_13-m1
#> ACTG1           6         0          4          7         14         13
#> ARHGDIB         3         0          2          2          2          2
#> BTF3            9         0          2          2          6          2
#> CD3D            2         0          4          1          2          0
#> CD3E            3         2          5          6          7          2
#> CFL1            2         0          1          3          1          7
#>         cell_14-m1 cell_15-m1 cell_16-m1 cell_17-m1 cell_18-m1 cell_19-m1
#> ACTG1            1          1          5          6          3          1
#> ARHGDIB          8          0          3          3          2          0
#> BTF3             3          2          5          1          1          4
#> CD3D             3          1          0          5          6          4
#> CD3E             1          4          2          2          4         12
#> CFL1             2          4          2          0          0          1
#>         cell_20-m1 cell_21-m1 cell_22-m1 cell_23-m1 cell_24-m1 cell_25-m1
#> ACTG1            7          4          5          2          1          6
#> ARHGDIB          2          5          1          2          3          1
#> BTF3             5          3          2          6          0          0
#> CD3D             0          2          4          2          0          1
#> CD3E             2          3          2          4          3          2
#> CFL1             3          1          2          0          1          3
#>         cell_26-m1 cell_27-m1 cell_28-m1 cell_29-m1 cell_30-m1 cell_31-m1
#> ACTG1            8         10          3          4          1          7
#> ARHGDIB          1          0          1          3          1          0
#> BTF3             6          0          3          1          2          6
#> CD3D             1          2          0          0          0          1
#> CD3E             8          6          0          0          1          5
#> CFL1             2          8          3          4          6          2
#>         cell_32-m1 cell_33-m1 cell_34-m1 cell_35-m1 cell_36-m1 cell_37-m1
#> ACTG1            3          8          5          2          6          3
#> ARHGDIB          3          8          0          1          1          0
#> BTF3             0          8          2          7          1          1
#> CD3D             2          0          0          1          2          3
#> CD3E             2          0          1          0          2          4
#> CFL1             1          1          0          3          0          0
#>         cell_38-m1 cell_39-m1 cell_40-m1 cell_1-m2 cell_2-m2 cell_3-m2
#> ACTG1            8          5          4         4         4         7
#> ARHGDIB          1          5          0         4         3         3
#> BTF3             2          3          0         2         2         2
#> CD3D             1          1          1         1         3        12
#> CD3E             3          3          1         0         3         1
#> CFL1            10          0          1         0         1         2
#>         cell_4-m2 cell_5-m2 cell_6-m2 cell_7-m2 cell_8-m2 cell_9-m2 cell_10-m2
#> ACTG1           1         1         1         5         2         4          9
#> ARHGDIB         0         2         4         1         5         1          0
#> BTF3            1         7         5         1         2         3          1
#> CD3D            1         2         2         0         3         4          1
#> CD3E            4         1         1         0         4         4          4
#> CFL1            1         0         2         4         2         1          6
#>         cell_11-m2 cell_12-m2 cell_13-m2 cell_14-m2 cell_15-m2 cell_16-m2
#> ACTG1            8          5          5          4          4          5
#> ARHGDIB          1          1          4          7          0          3
#> BTF3             2          3          3          2          5          0
#> CD3D             2          4          2         10          6          1
#> CD3E             0          7          3          3          8          2
#> CFL1             3          8          1          7          5          2
#>         cell_17-m2 cell_18-m2 cell_19-m2 cell_20-m2 cell_21-m2 cell_22-m2
#> ACTG1            8          4          1          6          3          4
#> ARHGDIB          0          8          1          0          0          2
#> BTF3             0          3          1          1          2          2
#> CD3D             0          2          1          6          2          2
#> CD3E             2          4          1          1          6          3
#> CFL1             3          2          2          2          0          3
#>         cell_23-m2 cell_24-m2 cell_25-m2 cell_26-m2 cell_27-m2 cell_28-m2
#> ACTG1            2          6          1          1          2          7
#> ARHGDIB          3          0          4          0          3          3
#> BTF3             2          0          0          1          2          8
#> CD3D             0          1          1          4          1          0
#> CD3E             2          7          3          5         14          1
#> CFL1             4          1          2          1          1          6
#>         cell_29-m2 cell_30-m2 cell_31-m2 cell_32-m2 cell_33-m2 cell_34-m2
#> ACTG1            6          2          6          1          6          1
#> ARHGDIB          1          2          0          3          4          3
#> BTF3             3          3          0          3          5          4
#> CD3D             2          7          1          1          2          0
#> CD3E            12          2          1          2          0          1
#> CFL1             3          0          0          2          2          3
#>         cell_35-m2 cell_36-m2 cell_37-m2 cell_38-m2 cell_39-m2 cell_40-m2
#> ACTG1           10          3          5          5          4          9
#> ARHGDIB          8          4          2          1          1          1
#> BTF3             0          2          1          5          1          2
#> CD3D             1          1          3          2          1          2
#> CD3E             1          1          6          1          0          3
#> CFL1             4          1          2          1          1          3
#>         cell_1-m3 cell_2-m3 cell_3-m3 cell_4-m3 cell_5-m3 cell_6-m3 cell_7-m3
#> ACTG1           4         3         4         4         3         2         8
#> ARHGDIB         5         5         0         8         1         5         0
#> BTF3            2         2         1         1         1         5         0
#> CD3D            4         0         3         7         1         0         4
#> CD3E            2         2         1         2         2         3         3
#> CFL1            4         1         2         0         3         2         2
#>         cell_8-m3 cell_9-m3 cell_10-m3 cell_11-m3 cell_12-m3 cell_13-m3
#> ACTG1           5         2         18          2          7          5
#> ARHGDIB         6         1          1          1          0          3
#> BTF3            3         0          5          2          5          1
#> CD3D            8         6          1          0          0          0
#> CD3E            9         4          4          0          4          3
#> CFL1            1         1         13          1          0          2
#>         cell_14-m3 cell_15-m3 cell_16-m3 cell_17-m3 cell_18-m3 cell_19-m3
#> ACTG1            2          7          5          7          3          1
#> ARHGDIB          3          6          3          1          0          1
#> BTF3             3          2          2          6          2          2
#> CD3D             2          2          4          9          6          4
#> CD3E             1          2          4          4          6          5
#> CFL1             1          4          2          9          2          4
#>         cell_20-m3 cell_21-m3 cell_22-m3 cell_23-m3 cell_24-m3 cell_25-m3
#> ACTG1            6          4          8          3          5          2
#> ARHGDIB          4          5          1          1          3          6
#> BTF3             0          1          1         11         10          1
#> CD3D             2          3          2          3          1          1
#> CD3E             3          2          1          9          5          2
#> CFL1             1          3          3          4          5          0
#>         cell_26-m3 cell_27-m3 cell_28-m3 cell_29-m3 cell_30-m3 cell_31-m3
#> ACTG1            7          1         14          5          9          8
#> ARHGDIB          0          4          2          6          1          3
#> BTF3             5          2          0          5          1          0
#> CD3D             4          1          2          0          6          3
#> CD3E             1          0          2          3          4          3
#> CFL1             0         10          3          4          3          5
#>         cell_32-m3 cell_33-m3 cell_34-m3 cell_35-m3 cell_36-m3 cell_37-m3
#> ACTG1            2          6          8          6          3          6
#> ARHGDIB          1          1          7          0          2          6
#> BTF3            11          1         10          2          9          0
#> CD3D             3          3          0          2          2          0
#> CD3E             2          1          7          3          2          1
#> CFL1             1          5          1          3          1          5
#>         cell_38-m3 cell_39-m3 cell_40-m3
#> ACTG1            3          1          6
#> ARHGDIB          0         11          0
#> BTF3             2          6          1
#> CD3D             1          2          0
#> CD3E             3          6          2
#> CFL1            10          5          0