Generate Adjacency Matrices from Gene Interaction Tables

Constructs adjacency matrices from a list of data frames, each representing weighted gene–gene interactions.

Usage

generate_adjacency(df_list, nCores = 1)

Arguments

df_list

A list of data frames. Each data frame must have three columns:

Gene1

Character. First gene in the interaction.

Gene2

Character. Second gene in the interaction.

Weight

Numeric. Weight or strength of the interaction from Gene1 to Gene2.

nCores

Integer. Number of CPU cores to use for parallel processing. Defaults to the number of available workers from the current BiocParallel backend.

Value

A list of square numeric adjacency matrices. Each matrix has genes as row and column names, and weights as values. Diagonal entries are set to zero (no self-interactions).

Details

The function first identifies all unique genes across all data frames to define the matrix dimensions. For each interaction table, it places the corresponding weights at the appropriate gene-pair positions. Parallelization is handled by BiocParallel for improved performance on multiple datasets.

Missing weights (NA) are ignored during construction. Only gene pairs matching the global gene list are inserted.

Examples

data("count_matrices")

networks <- infer_networks(
    count_matrices_list = count_matrices,
    method = "GENIE3",
    nCores = 1
)
head(networks[[1]])
#>   regulatoryGene targetGene    weight
#> 1          EIF3E      PCBP1 0.1490520
#> 2           CD3E       CD3D 0.1469628
#> 3          PTPRC       CYBA 0.1404093
#> 4          EEF1G      PTPRC 0.1347624
#> 5         SNRPD2       CD3D 0.1297852
#> 6         EEF1A1       NACA 0.1286103

wadj_list <- generate_adjacency(networks)
head(wadj_list[[1]])
#>               ACTG1    ARHGDIB       BTF3       CD3D        CD3E       CFL1
#> ACTG1   0.000000000 0.01498779 0.02246295 0.01021727 0.029352222 0.07908752
#> ARHGDIB 0.014001092 0.00000000 0.02241914 0.01013807 0.015683181 0.01272818
#> BTF3    0.028149458 0.04454537 0.00000000 0.01294270 0.025942891 0.01647918
#> CD3D    0.009271792 0.01124372 0.01087787 0.00000000 0.096221007 0.01792420
#> CD3E    0.051331393 0.02610803 0.04631266 0.14696284 0.000000000 0.02247259
#> CFL1    0.083894034 0.01159182 0.01607212 0.02181620 0.009070981 0.00000000
#>             COX4I1       COX7C        CYBA     EEF1A1      EEF1D      EEF1G
#> ACTG1   0.01974538 0.031066699 0.025419545 0.02681652 0.03417896 0.03313300
#> ARHGDIB 0.02011559 0.015339802 0.015193464 0.01205679 0.01124851 0.02493718
#> BTF3    0.01361029 0.009766704 0.025199552 0.03366413 0.02130550 0.01886889
#> CD3D    0.02786161 0.012647962 0.020022968 0.01980636 0.02809625 0.00980186
#> CD3E    0.01761759 0.039158148 0.019609928 0.01724225 0.01794807 0.03041487
#> CFL1    0.02566644 0.011924315 0.007945201 0.03747520 0.02026030 0.01328759
#>               EEF2       EIF1      EIF3E       FTH1        FTL      HLA-A
#> ACTG1   0.03346333 0.01798438 0.06117734 0.01997940 0.01499651 0.04175981
#> ARHGDIB 0.01638393 0.02586012 0.01037524 0.04303134 0.02892881 0.01550802
#> BTF3    0.01860454 0.03339785 0.01310516 0.01280956 0.01780043 0.02045866
#> CD3D    0.01879681 0.01426987 0.01328735 0.02909073 0.01292262 0.02136655
#> CD3E    0.01161500 0.02716048 0.04571429 0.06217546 0.01336658 0.03854652
#> CFL1    0.01146660 0.02311049 0.01168145 0.01985527 0.02465710 0.02142972
#>              HLA-B      HLA-C      HLA-E    HNRNPA1  HNRNPA2B1       MYL6
#> ACTG1   0.01926184 0.01961218 0.01727767 0.02054180 0.03219929 0.02956920
#> ARHGDIB 0.01452279 0.01552750 0.01882363 0.01392340 0.01762726 0.02020984
#> BTF3    0.02174968 0.01601926 0.05026469 0.01814759 0.04742720 0.03140695
#> CD3D    0.01219276 0.02298767 0.01586175 0.01188190 0.02974866 0.01223848
#> CD3E    0.02139216 0.02061048 0.01484055 0.02656332 0.01488804 0.04492836
#> CFL1    0.04305174 0.04648829 0.03130131 0.01374211 0.01759466 0.03512154
#>               NACA      PCBP1      PCBP2       PFN1      PTPRC    PTPRCAP
#> ACTG1   0.02543838 0.03436341 0.01386494 0.01389119 0.02197100 0.02882272
#> ARHGDIB 0.01669668 0.03126698 0.01770128 0.01235499 0.01422497 0.01308290
#> BTF3    0.03700249 0.02140567 0.04049582 0.02844841 0.01548292 0.05473062
#> CD3D    0.01856739 0.01177420 0.03483753 0.03686438 0.01473891 0.01561534
#> CD3E    0.06558703 0.01677658 0.02221745 0.01979102 0.01251589 0.01247935
#> CFL1    0.01744675 0.01938389 0.01392080 0.01589007 0.01238552 0.03319928
#>               RAC2       SKP1     SNRPD2     TMSB4X      UBA52        UBC
#> ACTG1   0.04105920 0.01849791 0.01672115 0.04502034 0.03943129 0.06293009
#> ARHGDIB 0.08479629 0.01924147 0.01421459 0.01320316 0.01973049 0.01618229
#> BTF3    0.04484631 0.02259783 0.07120196 0.01075059 0.01627598 0.02086301
#> CD3D    0.01452884 0.01270937 0.06203494 0.01028037 0.01148994 0.01684749
#> CD3E    0.01923977 0.10685161 0.02476148 0.01468582 0.06553350 0.02153051
#> CFL1    0.01659046 0.04037810 0.02214337 0.01836776 0.01130885 0.01949877
#>              UQCRB       YBX1
#> ACTG1   0.01828085 0.05325515
#> ARHGDIB 0.00978018 0.03666977
#> BTF3    0.01338937 0.05126435
#> CD3D    0.02492224 0.01536637
#> CD3E    0.02256771 0.02751116
#> CFL1    0.01978660 0.07748529