Simulate Zero-Inflated Negative Binomial (ZINB) Count Matrices with Sequencing Depth

Simulates one or more count matrices following a zero-inflated negative binomial (ZINB) distribution, incorporating gene-gene interaction structures and cell-specific sequencing depth variation.

Usage

zinb_simdata(
  n,
  p,
  B,
  mu_range,
  mu_noise,
  theta,
  pi,
  kmat = 1,
  depth_range = NA
)

Arguments

n

Integer. Number of cells (samples) in each simulated matrix.

p

Integer. Number of genes (features) in each simulated matrix.

B

A symmetric binary adjacency matrix (0/1) defining gene-gene connectivity. Row and column names correspond to gene names.

mu_range

List of numeric vectors (length 2 each). Range of gene expression means for each simulated matrix.

mu_noise

Numeric vector. Mean of background noise for each matrix.

theta

Numeric vector. Dispersion parameters of the negative binomial distribution for each matrix. Smaller theta implies higher overdispersion.

pi

Numeric vector. Probability of excess zeros (0 < pi < 1) for each matrix.

kmat

Integer. Number of count matrices to simulate. Default is 1.

depth_range

Numeric vector of length 2 or NA. Range of total sequencing depth per cell. If NA, no depth adjustment is performed.

Value

A list containing kmat matrices. Each matrix has:

Rows representing cells (cell_1, ..., cell_n).

Columns representing genes (rownames(B)).

Count values following a ZINB distribution.

Details

Each simulated matrix:

Generates gene expression values based on a ZINB model.

Modulates expression using the adjacency matrix B.

Applies random sequencing depth scaling if depth_range is provided.

Useful for benchmarking single-cell RNA-seq network inference methods with dropout events and network structure.

Examples

B <- matrix(0, nrow = 5, ncol = 5)
rownames(B) <- colnames(B) <- paste0("Gene", 1:5)
B[1, 2] <- B[2, 3] <- B[4, 5] <- 1
B <- B + t(B)
zinb_matrices <- zinb_simdata(
    n = 100, p = 5, B = B,
    mu_range = list(c(1, 5), c(2, 6)),
    mu_noise = c(0.5, 0.7),
    theta = c(1, 2),
    pi = c(0.2, 0.3),
    kmat = 2,
    depth_range = c(500, 5000)
)
print(zinb_matrices[[1]])
#>          Gene1 Gene2 Gene3 Gene4 Gene5
#> cell_1     700   280     0   140   140
#> cell_2       0   857  1999  1142   857
#> cell_3       0   465   116   697     0
#> cell_4     182   237    73    55    73
#> cell_5       0   448   448   448     0
#> cell_6     335   223     0    56   279
#> cell_7    1216   405     0  2026  1216
#> cell_8     106  1699   319  1274  1062
#> cell_9     304   708   708  1012   607
#> cell_10    262     0     0  2231   394
#> cell_11     67   267   100   267   400
#> cell_12    368   123   245   368   123
#> cell_13    330   412   495   247    82
#> cell_14    822  1028   822   617   617
#> cell_15    191   763  1145     0   954
#> cell_16    100     0   500   350   100
#> cell_17     55   165   385   551   771
#> cell_18    934   779   623     0   156
#> cell_19    134     0    67   268   234
#> cell_20   1199   600     0  1599     0
#> cell_21    657  1751  2189     0     0
#> cell_22    448  1121   448   673  1345
#> cell_23      0  2178   545   545  1089
#> cell_24    323  1293  1293   323   646
#> cell_25    199   798   399   399   399
#> cell_26    409   409  1228   819   819
#> cell_27      0   301   301  1202  1503
#> cell_28    173   520  1213   347  1213
#> cell_29    180   451   361   631   992
#> cell_30    236  1418  1182   945   473
#> cell_31    807   908   303  1110  1514
#> cell_32    347   397   496   595   297
#> cell_33      0   308   257   308  1283
#> cell_34      0   160    80   400   480
#> cell_35     96   305   241    32   128
#> cell_36    925   231   578   925   810
#> cell_37    195   160   160   195   195
#> cell_38   1548  1548     0   310     0
#> cell_39      0  1340   893   223   223
#> cell_40   2137   712   237   712   475
#> cell_41    347   347   278   347     0
#> cell_42    142    71   214   356   427
#> cell_43   1350  1350   386   386   386
#> cell_44   1064  1241   709   355     0
#> cell_45    181   723   181   181   542
#> cell_46    493     0   985     0   493
#> cell_47    446  1338   446  1338     0
#> cell_48   1131   905   679   905   226
#> cell_49    809   708  1213   809   202
#> cell_50    369   738  2398   553   369
#> cell_51     66   328   328   525   328
#> cell_52   1466  2053     0   293   293
#> cell_53     44   438   219   263   131
#> cell_54    140   245   140   245   420
#> cell_55      0   284   284     0   569
#> cell_56      0   331   331   331   372
#> cell_57    289  1155  1155   866  1010
#> cell_58    312   572    52   312   468
#> cell_59      0  1714   343   685   343
#> cell_60   1172  1758   146   293   146
#> cell_61    868   124    62  1364   372
#> cell_62    262   197   197   328   590
#> cell_63    140  1404   421   140   421
#> cell_64   1322   721     0   481   601
#> cell_65   1679   630   840   420     0
#> cell_66    825   550   275   275  2475
#> cell_67    329   658   987  1316   987
#> cell_68   1840  1380   460     0     0
#> cell_69   1230  1230   820   410   410
#> cell_70      0  2262     0  2262     0
#> cell_71      0   481  1924   481   481
#> cell_72    415   415   415   831   831
#> cell_73    157   471   157  1257   157
#> cell_74   1264   805   230   345   575
#> cell_75   1244  1244  1244   622     0
#> cell_76    145   218   435   290   145
#> cell_77      0     0  1016     0  2032
#> cell_78      0  3176     0     0   635
#> cell_79    362   181   181   724   543
#> cell_80    137   411   137     0     0
#> cell_81   1110   555   555  1665     0
#> cell_82    128   128   256   192   192
#> cell_83      0     0     0     0     0
#> cell_84      0  1695   847   847     0
#> cell_85     81   162   135   135   162
#> cell_86      0     0     0     0     0
#> cell_87    547  1094  1641   547   547
#> cell_88    284  2269   567   284     0
#> cell_89    419  2095     0     0     0
#> cell_90     94   235   281    94   235
#> cell_91    913     0     0  1141   228
#> cell_92      0     0  2798     0     0
#> cell_93    567   850   850   850   567
#> cell_94    235   469     0     0     0
#> cell_95      0     0     0  2387  2387
#> cell_96   1190  1190   397   397   397
#> cell_97    374  1122  1122  1496     0
#> cell_98    113   113   113   113   226
#> cell_99      0     0     0     0  3697
#> cell_100     0   806   806   806   806